Added ability to hide "End-to-End Attribute Flows" sections.
The "Hide Default Sync Rules" option selection now hides the OOB rules when not only the rule is not changed at all but also when only precedence number or tag is changed (due to upgrade / forest addition).

The visibility control checkboxes are now enabled only after the browser loads the report completely.